+++ .gitignore
@@ -1,7 +1,145 @@
-*.ipynb_checkpoints
-__pycache__
-*.ipynb
-tb/*
-results/*
-checkpoints/*
-actor-critic/*
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+
+# server scitps output
+*.std*
+slurm-*.out
+
+# C extensions
+*.so
+
+# Data
+data
+resources
+tests/*.ipynb
+notes.md
+
+# Experiments
+*experiments*
+
+# Logs
+logs
+
+# Figures
+figures
+
+# java extensions
+.classpath
+.factorypath
+.project
+.settings/
+.settings
+.vscode
+
+# checkpoints, results, logs, saves, data, images
+
+# Distribution / packaging
+.Python
+env/
+*env*
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+wheels/
+*.egg-info/
+.installed.cfg
+*.egg
+**/.idea
+.nfs*
+.idea
+.idea/
+.idea/*
+*.iml
+envs/
+
+# PyInstaller
+#  Usually these files are written by a python script from a template
+#  before PyInstaller builds the exe, so as to inject date/other infos into it.
+*.manifest
+*.spec
+
+# Installer logs
+pip-log.txt
+pip-delete-this-directory.txt
+
+# Unit test / coverage reports
+htmlcov/
+.tox/
+.coverage
+.coverage.*
+.cache
+nosetests.xml
+coverage.xml
+*.cover
+.hypothesis/
+.pytest_cache/
+outputs
+
+# Translations
+*.mo
+*.pot
+
+# Django stuff:
+*.log
+local_settings.py
+
+# Flask stuff:
+instance/
+.webassets-cache
+
+# Scrapy stuff:
+.scrapy
+
+# Sphinx documentation
+docs/_build/
+
+# PyBuilder
+target/
+
+# Jupyter Notebook
+.ipynb_checkpoints
+.DS_Store
+__pycache__/*
+*/notebooks/dev
+*/dev/*
+
+
+# pyenv
+.python-version
+
+# celery beat schedule file
+celerybeat-schedule
+
+# SageMath parsed files
+*.sage.py
+
+# dotenv
+.env
+
+# virtualenv
+.venv
+venv/
+ENV/
+
+# Spyder project settings
+.spyderproject
+.spyproject
+
+# Rope project settings
+.ropeproject
+
+# mkdocs documentation
+/site
+
+# mypy
+.mypy_cache/
