--- config.yaml
@@ -1,42 +1,0 @@
-Files:
-  checkpoints_dir: 'checkpoints/'
-  tensorboard_log_dir: 'tensorboard/'
-  results_log_path: 'results.log'
-
-Seed: 
-
-Agent:
-  name: 'actor-critic'
-  gamma: 1.0
-  Network:
-    name: 'policy-value'
-    use_bias: true
-    # bias_init_const: 1.0e-5
-    bias_init_const: 0.0008
-    grad_clip_norm: 20.0     # norm of gradients to clip
-    n_filters: 64
-    activation: 'elu'
-    state_embedding_size: 128
-    state_channels: 3
-    lr: 5.0e-5
-    actor_coeff: 0.01 # 1.0
-    critic_coeff: 100.0 #1.0e+7
-    reg_coeff: 1.0e-4
-    entropy_coeff: 0.0
-    value_activation: 'identity'
-
-Training:
-  init_checkpoint: 0      # 0 for training from scratch, any integer > 0 for restoring the model from the last checkpoint 
-  n_iter: 800
-  n_games: 16            
-  n_workers: 8
-  display_every: 5
-  batch_size: 64
-  buffer_size: 1024
-  T_update_net: 4
-  T_update_target_net: 50
-  prefill_buffer: true
-
-Evaluation:
-  n_games: 30
-  n_workers: 10
